+++ src/ops.rs
@@ -11,7 +11,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-use crate::core::{Args, AmbosoEnv, AmbosoMode, AmbosoLintMode, AnvilKern, INVIL_VERSION, INVIL_OS, EXPECTED_AMBOSO_API_LEVEL, parse_stego_toml, lex_stego_toml, SemVerKey, ANVIL_INTERPRETER_TAG_REGEX, RULE_REGEX, RULELINE_MARK_CHAR, RULEWARN_REGEX, cut_line_at_char, CutDirection, semver_compare, MIN_AMBOSO_V_PYKERN, MIN_AMBOSO_V_CHECK_DETACHED};
+use crate::core::{Args, AmbosoEnv, AmbosoMode, AmbosoLintMode, AnvilKern, INVIL_VERSION, INVIL_OS, EXPECTED_AMBOSO_API_LEVEL, parse_stego_toml, lex_stego_toml, SemVerKey, ANVIL_INTERPRETER_TAG_REGEX, RULE_REGEX, RULELINE_MARK_CHAR, RULEWARN_REGEX, cut_line_at_char, CutDirection, semver_compare, MIN_AMBOSO_V_PYKERN, MIN_AMBOSO_V_CHECK_DETACHED, MIN_AMBOSO_V_CUST_RECIPES};
 use crate::utils::try_parse_stego;
 
 use std::process::{Command, Stdio, exit};
@@ -28,7 +28,8 @@ use std::cmp::Ordering;
 
 #[cfg(feature = "anvilPy")]
 use crate::anvil_py::{ANVILPY_UNPACKDIR_NAME,unpack_srcdist, post_unpack};
-
+#[cfg(feature = "anvilCustom")]
+use crate::anvil_custom::find_anvilcustom_recipe;
 
 pub fn do_build(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
     match args.tag {
@@ -615,6 +616,71 @@ pub fn do_delete(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
         None => {
             warn!("No tag provided.");
             Err("No tag provided".to_string())
+        }
+    }
+}
+
+pub fn do_purge(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
+    match env.run_mode {
+        Some(ref runmode) => {
+            match runmode {
+                AmbosoMode::GitMode => {
+                    debug!("Doing purge for git mode");
+                    let mut args_copy = args.clone();
+                    let mut res = 0;
+                    for tag in env.gitmode_versions_table.keys() {
+                        args_copy.tag = Some(tag.to_string());
+                        let delete_res = do_delete(env,&args_copy);
+                        match delete_res {
+                            Ok(s) => {
+                                trace!("{}", s);
+                            }
+                            Err(e) => {
+                                warn!("do_purge(): Delete failed for tag {{{}}}. Err: {}", tag, e);
+                                res += 1;
+                            }
+                        }
+                    }
+                    if res == 0 {
+                        Ok("Purge success".to_string())
+                    } else {
+                        Err(format!("Purge error: {res} failures"))
+                    }
+                }
+                AmbosoMode::BaseMode => {
+                    debug!("Doing purge for base mode");
+                    let mut args_copy = args.clone();
+                    let mut res = 0;
+                    for tag in env.basemode_versions_table.keys() {
+                        args_copy.tag = Some(tag.to_string());
+                        let delete_res = do_delete(env,&args_copy);
+                        match delete_res {
+                            Ok(s) => {
+                                trace!("{}", s);
+                            }
+                            Err(e) => {
+                                warn!("do_purge(): Delete failed for tag {{{}}}. Err: {}", tag, e);
+                                res += 1;
+                            }
+                        }
+                    }
+                    if res == 0 {
+                        Ok("Purge success".to_string())
+                    } else {
+                        Err(format!("Purge error: {res} failures"))
+                    }
+                }
+                AmbosoMode::TestMode => {
+                    todo!("Purge op for test mode");
+                }
+                AmbosoMode::TestMacro => {
+                    todo!("Purge op for test macro mode");
+                }
+            }
+        }
+        None => {
+            error!("Invalid: None env.run_mode");
+            Err("Invalid None env.run_mode".to_string())
         }
     }
 }
@@ -1659,10 +1725,10 @@ pub fn lex_makefile(file_path: impl AsRef<Path>, dbg_print: bool, skip_recap: bo
 
 fn build_step(args: &Args, env: &AmbosoEnv, cflg_str: String, query: &str, bin_path: PathBuf, target_path: PathBuf, bin: String, head_was_detached: bool, do_postbuild: bool) -> Result<String,String> {
     let output;
-    let build_step_command;
+    let build_step_command: String;
     match env.anvil_kern {
         AnvilKern::AmbosoC => {
-            build_step_command = "make";
+            build_step_command = "make".to_string();
         }
         AnvilKern::AnvilPy => {
             let mut use_python_build = true;
@@ -1676,16 +1742,31 @@ fn build_step(args: &Args, env: &AmbosoEnv, cflg_str: String, query: &str, bin_p
                 }
             }
             if use_python_build {
-                build_step_command = "python";
+                build_step_command = "python".to_string();
             } else {
-                build_step_command = "make";
+                build_step_command = "make".to_string();
             }
         }
         AnvilKern::Custom => {
             #[cfg(feature = "anvilCustom")] {
                 match &env.anvilcustom_env {
                     Some(cust_env) => {
-                        build_step_command = &cust_env.custom_builder;
+                        match semver_compare(&env.anvil_version, MIN_AMBOSO_V_CUST_RECIPES) {
+                            Ordering::Less => {
+                                build_step_command = cust_env.custom_builder.clone();
+                            },
+                            _ => {
+                                match find_anvilcustom_recipe(&cust_env, query) {
+                                    Some(recipe) => {
+                                        build_step_command = recipe.build.clone();
+                                    },
+                                    None => {
+                                        error!("Could not find recipe for {}", query);
+                                        return Err(format!("Missing recipe for {}", query));
+                                    }
+                                };
+                            }
+                        }
                     }
                     None => {
                         error!("Missing anvilcustom_env");
@@ -1696,7 +1777,7 @@ fn build_step(args: &Args, env: &AmbosoEnv, cflg_str: String, query: &str, bin_p
             #[cfg(not(feature = "anvilCustom"))] {
                 // Handle AnvilCustom case when the feature is not enabled
                 error!("AnvilCustom kern feature is not enabled");
-                return Err("AnvilCustom kern feauture is not enabled".to_string());
+                return Err("AnvilCustom kern feature is not enabled".to_string());
             }
         }
     }
@@ -1704,7 +1785,7 @@ fn build_step(args: &Args, env: &AmbosoEnv, cflg_str: String, query: &str, bin_p
     match env.anvil_kern {
         AnvilKern::AmbosoC => {
             if args.no_rebuild {
-                let mut cmd = Command::new(build_step_command);
+                let mut cmd = Command::new(&build_step_command);
                 for arg in &args.extra_args {
                     cmd.arg(arg);
                 }
@@ -1726,7 +1807,7 @@ fn build_step(args: &Args, env: &AmbosoEnv, cflg_str: String, query: &str, bin_p
         }
         AnvilKern::AnvilPy => {
             debug!("Running \'{build_step_command}\'");
-            output = Command::new(build_step_command)
+            output = Command::new(&build_step_command)
                 .arg("-m")  // Using -o bin_path would allow skipping the mv command
                 .arg("build")
                 .output()
@@ -1734,21 +1815,148 @@ fn build_step(args: &Args, env: &AmbosoEnv, cflg_str: String, query: &str, bin_p
         }
         AnvilKern::Custom => {
             // "custom_builder" "target_d" "builds_dir" "bin_name" "q_tag" "stego_dir"
-            let mut cmd = Command::new(build_step_command);
+            let mut cmd = Command::new(&build_step_command);
 
-            cmd.arg(target_path.clone())
-            .arg(env.builds_dir.clone().expect("failed initialising builds_dir"))
-            .arg(bin.clone())
-            .arg(query)
-            .arg(env.stego_dir.clone().expect("failed initialising stego_dir"));
+            match semver_compare(&env.anvil_version, MIN_AMBOSO_V_CUST_RECIPES) {
+                Ordering::Less => {
+                    cmd.arg(target_path.clone())
+                    .arg(env.builds_dir.clone().expect("failed initialising builds_dir"))
+                    .arg(bin.clone())
+                    .arg(query)
+                    .arg(env.stego_dir.clone().expect("failed initialising stego_dir"));
+                },
+                _ => {
+                    match &env.anvilcustom_env {
+                        Some(cust_env) => {
+                            match find_anvilcustom_recipe(&cust_env, query) {
+                                Some(recipe) => {
+                                    if let Some(ref conf) = recipe.conf {
+                                        let conf_path = Path::new(&conf);
+                                        if !conf_path.exists() && recipe.prep.is_some() {
+                                            debug!("Running prep {}", recipe.prep.as_ref().unwrap());
+                                            let mut prep_cmd = Command::new(&recipe.prep.as_ref().unwrap());
+                                            let prep_output = prep_cmd.output().expect("failed to execute process");
+                                            match prep_output.status.code() {
+                                                Some(prep_ec) => {
+                                                    if prep_ec == 0 {
+                                                        debug!("{{{}}} succeded with status: {}", recipe.prep.as_ref().unwrap(), prep_ec.to_string());
+                                                    } else {
+                                                        warn!("{{{}}} failed with status: {}", recipe.prep.as_ref().unwrap(), prep_ec.to_string());
+                                                        io::stdout().write_all(&prep_output.stdout).unwrap();
+                                                        io::stderr().write_all(&prep_output.stderr).unwrap();
+                                                        return Err(format!("{{{}}} failed", recipe.prep.unwrap()));
+                                                    }
+                                                },
+                                                None => {
+                                                    error!("{{{}}} command failed", recipe.prep.as_ref().unwrap());
+                                                    io::stdout().write_all(&prep_output.stdout).unwrap();
+                                                    io::stderr().write_all(&prep_output.stderr).unwrap();
+                                                    return Err(format!("{{{}}} command failed", recipe.prep.unwrap()));
+                                                }
+                                            }
+                                            if !conf_path.exists() {
+                                                error!("Failed prep for configurer: {{{}}}", conf_path.display());
+                                                return Err(format!("Failed prep for configurer: {{{}}}", conf_path.display()));
+                                            }
+                                        } else if !conf_path.exists() {
+                                            error!("Custom configurer does not exist: {}", conf);
+                                            return Err(format!("Custom configurer does not exist: {}", conf));
+                                        }
+                                        let mut conf_cmd = Command::new(&conf);
+                                        match &args.config {
+                                            Some(config_arg) => {
+                                                if let Some(split_config_arg) = shlex::split(config_arg) {
+                                                    conf_cmd.args(split_config_arg);
+                                                } else {
+                                                    error!("Invalid shell syntax in {{{config_arg}}}");
+                                                    return Err(format!("Invalid shell syntax: {{{config_arg}}}"));
+                                                }
+                                            }
+                                            None => {}
+                                        }
+                                        let conf_output = conf_cmd.output().expect("failed to execute process");
+                                        match conf_output.status.code() {
+                                            Some(conf_ec) => {
+                                                if conf_ec == 0 {
+                                                    debug!("{{{}}} succeded with status: {}", conf, conf_ec.to_string());
+                                                } else {
+                                                    warn!("{{{}}} failed with status: {}", conf, conf_ec.to_string());
+                                                    io::stdout().write_all(&conf_output.stdout).unwrap();
+                                                    io::stderr().write_all(&conf_output.stderr).unwrap();
+                                                    return Err(format!("{{{conf}}} failed"));
+                                                }
+                                            }
+                                            None => {
+                                                error!("{{{}}} command failed", conf);
+                                                io::stdout().write_all(&conf_output.stdout).unwrap();
+                                                io::stderr().write_all(&conf_output.stderr).unwrap();
+                                                return Err(format!("{{{conf}}} command failed"));
+                                            }
+                                        }
+                                    } else {
+                                        cmd.arg(target_path.clone())
+                                        .arg(env.builds_dir.clone().expect("failed initialising builds_dir"))
+                                        .arg(bin.clone())
+                                        .arg(query)
+                                        .arg(env.stego_dir.clone().expect("failed initialising stego_dir"));
+                                    }
+                                    let build_path = Path::new(&build_step_command);
+                                    if !build_path.exists() && recipe.prep.is_some() && recipe.conf.is_none() {
+                                        debug!("Running prep {}", recipe.prep.as_ref().unwrap());
+                                        let mut prep_cmd = Command::new(&recipe.prep.as_ref().unwrap());
+                                        let prep_output = prep_cmd.output().expect("failed to execute process");
+                                        match prep_output.status.code() {
+                                            Some(prep_ec) => {
+                                                if prep_ec == 0 {
+                                                    debug!("{{{}}} succeded with status: {}", recipe.prep.unwrap(), prep_ec.to_string());
+                                                } else {
+                                                    warn!("{{{}}} failed with status: {}", recipe.prep.as_ref().unwrap(), prep_ec.to_string());
+                                                    io::stdout().write_all(&prep_output.stdout).unwrap();
+                                                    io::stderr().write_all(&prep_output.stderr).unwrap();
+                                                    return Err(format!("{{{}}} failed", recipe.prep.unwrap()));
+                                                }
+                                            },
+                                            None => {
+                                                error!("{{{}}} command failed", recipe.prep.as_ref().unwrap());
+                                                io::stdout().write_all(&prep_output.stdout).unwrap();
+                                                io::stderr().write_all(&prep_output.stderr).unwrap();
+                                                return Err(format!("{{{}}} command failed", recipe.prep.unwrap()));
+                                            }
+                                        }
+                                        if !build_path.exists() {
+                                            error!("Failed prep for builder: {{{}}}", build_step_command);
+                                            return Err(format!("Failed prep for builder: {{{}}}", build_step_command));
+                                        }
+                                    }
+                                },
+                                None => {
+                                    error!("Could not find recipe for {}", query);
+                                    return Err(format!("Missing recipe for {}", query));
+                                }
+                            };
+                        }
+                        None => {
+                            error!("Missing anvilcustom_env");
+                            return Err("Missing anvilcustom_env".to_string());
+                        }
+                    }
+                }
+            }
 
             for arg in &args.extra_args {
                 cmd.arg(arg);
             }
 
             debug!("Running \'{:?}\'", cmd);
-            output = cmd.output()
-                .expect("failed to execute process");
+            match cmd.output() {
+                Ok(out) => {
+                    output = out;
+                }
+                Err(e) => {
+                    error!("Could not run {{{:?}}}: {e}", cmd);
+                    return Err(format!("Error on running {{{:?}}}: {e}", cmd));
+                }
+            }
         }
     }
     match output.status.code() {