+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.2.34] - Unreleased
+
+### Added
+
+- Add support for anvil_recipe with custom kern
+- Add support for stego, cgen, delete, run, purge subcommands
+
+### Changed
+
+- Collect purge op into op_purge()
+- Bump deps
+
 ## [0.2.33] - 2026-08-24
 
 ### Changed